requirements.txt
openai
boto3
awscli
python.dotenv